refactor: update stubs test
