+++ test/stubs/encrypter-decrypter-stub.ts
@@ -1,6 +1,6 @@
 import { Either, success } from '@/shared';
-import { UnauthorizedError } from '@/use-cases/errors';
-import { IEncrypter, IDecrypter, IPayload } from '@/use-cases/interfaces';
+import { UnauthorizedError } from '@/use-cases/user/errors';
+import { IEncrypter, IDecrypter, IPayload } from '@/use-cases/user/interfaces';
 
 export const makeEncrypter = (): IEncrypter => {
   class EncrypterStub implements IEncrypter {
