+++ test/stubs/use-cases-stub.ts
@@ -1,7 +1,8 @@
-import { SignUpUseCase } from '@/use-cases/sign-up';
-import { SignInUseCase } from '@/use-cases/sign-in';
-import { AuthenticationUseCase } from '@/use-cases/authentication';
-import { UpdateUserUseCase } from '@/use-cases/update-user';
+import { SignUpUseCase } from '@/use-cases/user/sign-up';
+import { SignInUseCase } from '@/use-cases/user/sign-in';
+import { AuthenticationUseCase } from '@/use-cases/user/authentication';
+import { UpdateUserUseCase } from '@/use-cases/user/update-user';
+import { ExternalSignInUseCase } from '@/use-cases/user/external-sign-in';
 import {
   makeUserRepository,
   makeHasher,
@@ -10,7 +11,6 @@ import {
   makeHashCompare,
   makeDecrypter,
 } from '@/test/stubs';
-import { ExternalSignInUseCase } from '@/use-cases/external-sign-in';
 
 export const makeSignUpUseCase = (): SignUpUseCase => {
   const userRepository = makeUserRepository();
