+++ test/stubs/validation-stub.ts
@@ -1,4 +1,4 @@
-import { IValidation } from '@/adapters/interfaces';
+import { IValidation } from '@/adapters/controllers/interfaces';
 
 export const makeValidation = (): IValidation => {
   class ValidationStub implements IValidation {
